requirements/requirements-libs.txt
Celery>=4.3
PyYAML==5.3.1
django-object-actions==3.0.1
